projetoverzel/requirements.txt
alembic==1.7.4
apparmor==3.0.3
appdirs==1.4.4
application-utility==1.3.2
asgiref==3.4.1
attrs==21.3.0
autopep8==1.6.0
Babel==2.9.1
bcrypt==3.2.0
Beaker==1.11.0
beautifulsoup4==4.10.0
blinker==1.4
Brotli==1.0.9
btrfsutil==5.15.1
CacheControl==0.12.6
cachetools==4.2.4
caffeine-ng==3.5.1.post27+g1e5febf.d20211213
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
configobj==5.1.0.dev0
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.1.1
cryptography==36.0.1
decorator==5.1.0
distlib==0.3.3
distro==1.6.0
Django==4.0.1
django-cors-headers==3.10.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
docopt==0.6.2
drf-spectacular==0.21.1
duplicity==0.8.21
email-validator==1.1.3
evdev==1.4.0
ewmh==0.1.6
extras==1.0.0
fasteners==0.16.3
filelock==3.4.0
fixtures==3.0.0
Flask==2.0.2
Flask-BabelEx==0.9.4
Flask-Compress==1.8.0
Flask-Gravatar==0.5.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.0.0
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security-Too==4.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
future==0.18.2
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==1.1.2
gssapi==1.6.14
gufw==21.4.0
html5lib==1.1
httplib2==0.20.3
idna==3.3
importlib-metadata==4.8.1
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jsonschema==4.3.3
layoutswitcherlib==0.8.22
ldap3==2.9
lensfun==0.3.95
LibAppArmor==3.0.3
libtorrent===1.2.15-build-libtorrent-rasterbar-src-libtorrent-rasterbar-1.2.15-bindings-python
lit==13.0.0.dev0
louis==3.20.0
lutris==0.5.9.1
lxml==4.6.5
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
meson==0.60.3
mock==3.0.5
more-itertools==8.10.0
msgpack==1.0.3
netsnmp-python==1.0a1
nose==1.3.7
npyscreen==4.10.5
oauth2client==4.1.3
openapi-codec==1.3.2
ordered-set==4.0.2
packaging==20.9
pacman-mirrors==4.23.2
paramiko==2.7.2
passlib==1.7.4
pbr==5.8.0
pep517==0.12.0
Pillow==8.4.0
platformdirs==2.4.1
ply==3.11
progress==1.6
protobuf==3.19.1
psutil==5.8.0
psycopg2==2.9.3
pulsectl==21.10.5
pwquality==1.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.21
PyDrive==1.3.1
PyGObject==3.42.0
PyJWT==2.3.0
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.0
pyrsistent==0.18.0
PySide6==6.2.2
pysmbc==1.0.23
python-dateutil==2.8.2
python-editor==1.0.4
python-xlib==0.31
pytz==2021.3
pyxdg==0.27
PyYAML==5.4.1
rados==2.0.0
ranger-fm==1.9.3
rbd==2.0.0
reportlab==3.6.3
requests==2.26.0
requests-file==1.5.1
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
rsa==4.8
setproctitle==1.2.2
shiboken2==5.15.2
shiboken6==6.2.2
simplejson==3.17.6
six==1.16.0
smbus==1.1
soupsieve==2.2.1
speaklater==1.3
SQLAlchemy==1.3.24
sqlparse==0.4.2
sshtunnel==0.4.0
TBB==0.2
team==1.0
termcolor==1.1.0
testtools==2.5.0
tldextract==3.1.2
toml==0.10.2
tomli==1.2.2
ueberzug==18.1.9
ufw==0.36.1
uritemplate==3.0.1
urllib3==1.26.7
validate==5.1.0.dev0
virtualenv==20.11.0
Wand==0.6.7
webencodings==0.5.1
Werkzeug==2.0.2
WTForms==2.3.3
wxPython==4.0.7.post2
zipp==3.6.0
